chore: rename survey extension utils and fix logging
