+++ src/entrypoints/surveys.ts
@@ -1,7 +1,7 @@
 import { generateSurveys } from '../extensions/surveys'
 
+import { canActivateRepeatedly } from '../extensions/surveys/surveys-extension-utils'
 import { assignableWindow } from '../utils/globals'
-import { canActivateRepeatedly } from '../extensions/surveys/surveys-utils'
 
 assignableWindow.__PosthogExtensions__ = assignableWindow.__PosthogExtensions__ || {}
 assignableWindow.__PosthogExtensions__.canActivateRepeatedly = canActivateRepeatedly
