+++ src/extensions/surveys.tsx
@@ -1,5 +1,4 @@
 import { PostHog } from '../posthog-core'
-import { doesSurveyUrlMatch } from '../posthog-surveys'
 import {
     Survey,
     SurveyAppearance,
@@ -16,7 +15,7 @@ import { addEventListener } from '../utils'
 import * as Preact from 'preact'
 import { useContext, useEffect, useMemo, useRef, useState } from 'preact/hooks'
 import { document as _document, window as _window } from '../utils/globals'
-import { createLogger } from '../utils/logger'
+import { doesSurveyUrlMatch, SURVEY_LOGGER as logger } from '../utils/survey-utils'
 import { isNull, isNumber } from '../utils/type-utils'
 import { createWidgetShadow, createWidgetStyle } from './surveys-widget'
 import { ConfirmationMessage } from './surveys/components/ConfirmationMessage'
@@ -40,9 +39,8 @@ import {
     style,
     SURVEY_DEFAULT_Z_INDEX,
     SurveyContext,
-} from './surveys/surveys-utils'
+} from './surveys/surveys-extension-utils'
 import { prepareStylesheet } from './utils/stylesheet-loader'
-const logger = createLogger('[Surveys]')
 
 // We cast the types here which is dangerous but protected by the top level generateSurveys call
 const window = _window as Window & typeof globalThis
