+++ src/extensions/surveys/components/ConfirmationMessage.tsx
@@ -1,11 +1,15 @@
 import { h } from 'preact'
 import { SurveyAppearance, SurveyQuestionDescriptionContentType } from '../../../posthog-surveys-types'
-import { defaultSurveyAppearance, getContrastingTextColor, renderChildrenAsTextOrHtml } from '../surveys-utils'
+import {
+    defaultSurveyAppearance,
+    getContrastingTextColor,
+    renderChildrenAsTextOrHtml,
+} from '../surveys-extension-utils'
 import { BottomSection } from './BottomSection'
 import { Cancel } from './QuestionHeader'
 
 import { useContext } from 'preact/hooks'
-import { SurveyContext } from '../../surveys/surveys-utils'
+import { SurveyContext } from '../surveys-extension-utils'
 
 export function ConfirmationMessage({
     header,
