+++ src/extensions/surveys/components/QuestionHeader.tsx
@@ -2,7 +2,7 @@ import { h } from 'preact'
 import { useContext } from 'preact/hooks'
 import { SurveyQuestionDescriptionContentType } from '../../../posthog-surveys-types'
 import { cancelSVG } from '../icons'
-import { SurveyContext, defaultSurveyAppearance, renderChildrenAsTextOrHtml } from '../surveys-utils'
+import { SurveyContext, defaultSurveyAppearance, renderChildrenAsTextOrHtml } from '../surveys-extension-utils'
 
 export function QuestionHeader({
     question,
